Fix tests and lint (with upgrade)
