+++ .github/workflows/lint.yaml
@@ -12,10 +12,10 @@ jobs:
     runs-on: ubuntu-latest
 
     steps:
-    - name: Check out code
-      uses: actions/checkout@v2
+      - name: Check out code
+        uses: actions/checkout@v2
 
-    - name: Lint
-      uses: golangci/golangci-lint-action@v2
-      with:
-        version: v1.42
+      - name: Lint
+        uses: golangci/golangci-lint-action@v2
+        with:
+          version: v1.45
